requirements.txt

mysql-connector-python
ipdb
ipython